Initial code generation
This commit introduces the Micronaut AOT code generators.
Code generators are meant to be called at build time, typically
via a build plugin.

They must be fed with a variety of configuration options, most
importantly the full runtime classpath of the application.
Each of the optimizers will then generate some code, which
must be understood in the larger sense:

- either snippets of Java code, typically static initializers
- or full Java classes
- or resources files (e.g service descriptors)
- or a combination of all the above

Special attention must be taken for the `OptimizedEntryPoint`
generator, which is the main source generator (it drives all
other generators). This source generator is responsible for
generating a new Micronaut application entry point for an
application.

For example, if the user has an application `demo.Application`,
this code generator will generate a new class called
`demo.Application$Optimized` (or any other name), which will
do a number of things:

- perform static initialization by calling the static initializers
generated by other source generators
- delegate to the user entry point

In practice, most source generators will end up calling the
Micronaut `StaticOptimizations.set(...)` method, which can be
used to inject optimizations at execution time.

Source generators fall into different categories:

1. pure algorithmical optimizations. The `EnvironmentProperties`
source generator is an example of that: it precomputes at build
time the combination of Micronaut properties from environment
variable names. This avoids performing the computation when the
application starts.
2. alternatives for expensive operations. The service loading
generators are such an example. In this case we would scan the
classpath at build time, look for beans which need to be instantiated,
eliminate those which do not meet the requirements, and generate
optimized service loaders.
3. substitutions. There are 2 different kinds of substitutions:
   - resource substiution: the example here is replacing `logback.xml`
with a Java based configuration
   - service substitution: replace the YAML properties loader with
a property source (therefore of a different service type) with precomputed
values

In the end, all those optimizations can be combined, enabled or disabled
separately. Note that some of them are dependent from each other.