Rework how the AOT module is configured
The AOT module now uses service loading to load optimizations.
It means that optimizations can be added just by registering
new services.

An optimization plugin must provide an ID, which is used to
enable or disable it. Then it requires a set of options, which
need to be found in a properties file.

Before, each optimization required a specific API, which made
it brittle with regards to change. Now each optimization is
self documenting, and the presence or absence of an optimization
configured in a properties file will not break the integration.

This also makes it much easier to integrate with third party
build plugins.