Add optimization for deduce environment
This commit introduces a new optimization: if `deduceEnvironment(true)`
is set, then we will get the result of the analysis, which is the list
of environment names and packages, and inject them directly in a
generated application context configurer, which itself disables
`deduceEnvironment`.

This means that the deployment environment must be the same as the
build environment.