Fix priority of generated property sources
The priority of the generated property sources (after conversion from Yaml)
wasn't low enough, preventing overriding from environment variables.
This commit adds an override of the `getOrder` method, which defaults
to the highest precedence, but it is possible to override the priority
of each generated property source by name.

Fixes #33