Add an option to configure the list of possible environments (#37)
This commit introduces a new configuration option (possible.environments)
which lists the names of environments which can be seen. This makes it
possible to load environments for optimized applications too, as long
as the bean requirements do not check for specific environments.

Without this, the only environments which are "seen" by AOT are the
environments which are active when analyzing the context. It means that
it cannot, for example, convert to yaml configuration files which are
for other environments.

By default, if you specify nothing, then the optimizations ARE using
the active environments. If you do specify the option, then it means
that you want to enable the user to specify different environments
at runtime, for example using the `MICRONAUT_ENVIRONMENTS` variable,
even for optimized applications.

This will require a change to Micronaut Core, which currently loads
_all_ configured environments statically as soon as a file is
converted to a constant property source (which is arguably a bug).

Closes #36