Add an option to "force" a particular environment (#46)
Currently the AOT analysis can be configured to tell the list of
_possible_ environments, for example so that we can convert _all_
configuration files to Java and not just the currently detected
environment. However, it doesn't let the user specify a particular
target environment for analysis. This may be useful when the
environement where the analysis is performed is known to be different
from the deployment environment. In particular, because the analysis
will be done based on bean requirements, having environments
forcefully activated can enable some beans which would otherwise
have been eliminated from the analysis.

To do this the user needs to configure the following property:

`target.environments=lambda`