Convert logback XML files to Java
This commit introduces proper conversion of logback.xml files to
Java. For this purpose, it uses logback 1.3. Because it's still
in alpha, this isn't great. However, the converter can use logback
1.3, while the actual application can use 1.2.x. This requires,
on the Gradle plugin side, to use:

```gradle
dependencies {
    aotPlugins("ch.qos.logback:logback-classic:1.3.0-alpha12")
}
```

It's also worth noting that the conversion implies the use
of logback classic.