Remove optimizer classpath
In practice, the AOT optimizer classpath and the application classpath
are mixed. The fact of putting the AOT optimizer classpath in a separate
argument creates confusion and can cause problems (if the Micronaut
version of AOT is different from the one under test).

Therefore, this commit removes the optimizer classpath option and uses
a single option.