Fix classloader isolation
This commit fixes how the CLI creates an isolated classloader:
before it wasn't possible to call the CLI without putting the
whole application classpath on the CLI classpath, which defeats
the purpose of the `--classpath` flag.

Now, the CLI can properly be called with just Micronaut AOT on
classpath, then the `--classpath` flag must be provided with
the union of the AOT classpath and the application classpath.