Fix deduceEnvironment optimization
This commit fixed the way the deduce environment optimization detects
if it should run or not. Previously it would only run if `deduceEnvironment`
was explicitly set to `true`, not if it was using the default.
With this change, if the user didn't say anything or if it's set to true,
then the optimization will run. Otherwise, we do nothing as it will
already be configured to `false`.