Introduce Netty optimization
This commit introduces a new optimization (id `netty.properties`) which
allows setting the `io.netty.machineId` and `io.netty.processId` system
properties before Netty reads them. Those system properties are used by
Netty in a composite key which is used to compute a channel id.

By default, if the system properties are not set, Netty will compute them,
which involves scanning MAC addresses or using JMX to get the current process
PID, which is expensive (up to 20ms on my machine).

This optimization, if enabled, overrides the system properties at startup,
by generating _random values at application startup_ for both the MAC
address and the PID. This is the default behavior of Netty when it cannot,
for some reason, infer those, so it should be safe.

Alternatively, the optimization allows for setting _constant_ values, which
is probably not recommended unless the distribution binary is bound to
a particular machine.