Add a BOM
This commit adds a BOM module for consistency with other Micronaut
modules. It is worth noting, however, that the BOM is unlikely to
be added in any user build, since AOT is a special module which is
integrated by build tools.

This commit also upgrades to the latest build plugins and enables
standardized project names.

Fixes #179